refactor: move app etc into new Start folder
