+++ src/AStar.Dev.OneDrive.Sync.Client/Start/App.axaml
@@ -1,7 +1,7 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:converters="using:AStar.Dev.OneDrive.Sync.Client.Converters"
-             x:Class="AStar.Dev.OneDrive.Sync.Client.App"
+             x:Class="AStar.Dev.OneDrive.Sync.Client.Start.App"
              RequestedThemeVariant="Default">
 
     <Application.Styles>
