+++ src/AStar.Dev.OneDrive.Sync.Client/Start/ApplicationServiceExtensions.cs
