+++ src/AStar.Dev.OneDrive.Sync.Client/Start/AuthenticationExtensions.cs
