+++ src/AStar.Dev.OneDrive.Sync.Client/Start/HttpClientExtension.cs
