+++ src/AStar.Dev.OneDrive.Sync.Client/Start/ViewModelExtensions.cs
